requirements.txt
aiogram==3.4
requests
python-dotenv
